Ready for release
Reverted version # back to v1.0 for launch, removed Footer for now as it
served no purpose, it will make it's return shortly. Got some basic
tables in, for now they will scroll horizontally on smaller devices, but
restructuring them for mobile is a good idea. Also anaged to save some
space with the typography stuff. Generated a minified css for those who
would use it. The template was tweaked a bit. Throwing the favicon up
here as well.